requirements.txt
google-search-results==2.4.2
python-dotenv==1.0.0
flask==3.0.2
pandas==2.2.1
openpyxl==3.1.2
gunicorn==21.2.0 